tests/wpt/tests/css/css-flexbox/parsing/flex-shrink-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Flexible Box Layout: getComputedStyle().flexShrink</title>
<link rel="help" href="https://drafts.csswg.org/css-flexbox/#propdef-flex-shrink">
<meta name="assert" content="flex-shrink computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("flex-shrink", "1");
test_computed_value("flex-shrink", "2.34e+06", ["2.34e+06", "2340000"]);
test_computed_value("flex-shrink", "6.78e+08", ["6.78e+08", "678000000"]);
test_computed_value("flex-shrink", "0");
</script>
</body>
</html>
