LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/parsing/flex-shrink-computed-expected.txt


PASS Property flex-shrink value '1' 
FAIL Property flex-shrink value '2.34e+06' assert_equals: expected "2.34e+06" but got "2340000"
FAIL Property flex-shrink value '6.78e+08' assert_equals: expected "6.78e+08" but got "678000000"
PASS Property flex-shrink value '0' 
